// old src/types/options.d.ts
export interface MessageArgumentOptions {
    type: MessageCommandArgsTypes;
    content?: Array<string>;
    match?: string;
    name: string;
    prompt?: {
        reply: string;
        send: string;
    };
    default?: string | Function;
}
export interface SlashArgumentOptions {
    type: 'string' | 'number' | 'boolean' | 'user' | 'channel' | 'role' | 'mentionable' | 'integer';
    name: string;
    choices: Array<{
        name: string;
        value: string;
    }>;
    required: boolean;
    description: string;
}
// ...
export interface InhibitorOptions {
    name: string;
    type: InhibitorType;
}